+++ Network/API/Telegram/Bot/Object/Sender.hs
@@ -18,24 +18,24 @@ import Network.API.Telegram.Bot.Property.Identifiable (Identifiable (Identificat
 
 data Sender
 	= Bot Int (Maybe Text) Text (Maybe Text) (Maybe Text)
-	| User Int (Maybe Text) Text (Maybe Text) (Maybe Text)
+	| Human Int (Maybe Text) Text (Maybe Text) (Maybe Text)
 	deriving Show
 
 instance Eq Sender where
 	Bot i _ _ _ _ == Bot i' _ _ _ _ = i == i'
-	User i _ _ _ _ == User i' _ _ _ _ = i == i'
+	Human i _ _ _ _ == Human i' _ _ _ _ = i == i'
 	_ == _ = False
 
 instance Identifiable Sender where
 	type Identificator Sender = Int
 	ident (Bot i _ _ _ _) = i
-	ident (User i _ _ _ _) = i
+	ident (Human i _ _ _ _) = i
 
 type Whom = Int -> Maybe Text -> Text -> Maybe Text -> Maybe Text -> Sender
 
 instance FromJSON Sender where
 	parseJSON = withObject "Sender" $ \v -> v .: "is_bot"
-		>>= bool (sender v User) (sender v Bot) where
+		>>= bool (sender v Human) (sender v Bot) where
 
 		sender :: Object -> Whom -> Parser Sender
 		sender v f = f <$> v .: "id" <*> v .:? "username"
@@ -44,12 +44,12 @@ instance FromJSON Sender where
 
 nickname :: Lens' Sender (Maybe Text)
 nickname f (Bot uid nn fn ln lng) = (\nn' -> Bot uid nn' fn ln lng) <$> f nn
-nickname f (User uid nn fn ln lng) = (\nn' -> User uid nn' fn ln lng) <$> f nn
+nickname f (Human uid nn fn ln lng) = (\nn' -> Human uid nn' fn ln lng) <$> f nn
 
 firstname :: Lens' Sender Text
 firstname f (Bot uid nn fn ln lng) = (\fn' -> Bot uid nn fn' ln lng) <$> f fn
-firstname f (User uid nn fn ln lng) = (\fn' -> User uid nn fn' ln lng) <$> f fn
+firstname f (Human uid nn fn ln lng) = (\fn' -> Human uid nn fn' ln lng) <$> f fn
 
 lastname :: Lens' Sender (Maybe Text)
 lastname f (Bot uid nn fn ln lng) = (\ln' -> Bot uid nn fn ln' lng) <$> f ln
-lastname f (User uid nn fn ln lng) = (\ln' -> User uid nn fn ln' lng) <$> f ln
+lastname f (Human uid nn fn ln lng) = (\ln' -> Human uid nn fn ln' lng) <$> f ln