+++ CHANGELOG.md
@@ -124,4 +124,5 @@
 * Introduce `Returning` type family to automatically decode the API response
 * Move `Poll` status info from `Poll` datatype to `Polling` constructor of `Content` datatype
 * Define `Persistable` instance for sending `Photo`
-* Separate `Venue` into module and datatype from `Info` datatype
+* Separate `Venue` and `Contact` into module and datatype from `Info` datatype
+* Rename `User` constructor of `Sender` datatype to `Human`